requirements.txt
# --- Core Framework ---
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0

# --- AI Models (LLM & Embeddings) ---
langchain-google-genai      # For Gemini 1.5 Flash
langchain-huggingface       # For BGE-M3 Embeddings
huggingface_hub             # To download models

# --- Database ---
faiss-cpu                 # The local vector DB engine

# --- File Handling (OCR & PDF) ---
pymupdf                     # Fast PDF reading
pdf2image                   # Converts PDF pages to images (for OCR)
pillow                      # Image processing
pytesseract                 # Fallback OCR

# --- Environment & UI ---
python-dotenv               # Loads .env file
streamlit                   # The Frontend UI
watchdog                    # Helps Streamlit reload faster
